Turn class component into functional component.
